+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 ## Unreleased
 - Adds [`Objectware.isOwnEmpty`][isOwnEmpty].
 - Adds [`Objectware.has`][has].
+- Adds [`Objectware.hasOwn`][hasOwn].
 
 [isOwnEmpty]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.isOwnEmpty
 [has]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.has
+[hasOwn]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.hasOwn
 
 ## 1.4.0 (Feb 8, 2015)
 - Adds [`Objectware.reject`][reject].